# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Films API
  version: 1.0.0
  description: API для управления фильмами и актёрами
paths:
  /movies:
    get:
      summary: Get all movies
      description: Возвращает список всех фильмов
      tags:
        - movies
      responses:
        '200':
          description: Список фильмов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Add a new movie
      description: Создаёт новый фильм
      tags:
        - movies
      requestBody:
        description: Данные нового фильма
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMovie'
      responses:
        '201':
          description: Фильм успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
  /movies/{id}:
    get:
      summary: Get movie by ID
      description: Возвращает информацию о фильме по ID
      tags:
        - movies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о фильме
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Фильм не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a movie by ID
      description: Обновляет данные фильма
      tags:
        - movies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Новые данные фильма
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMovie'
      responses:
        '200':
          description: Фильм успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Фильм не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a movie by ID
      description: Удаляет фильм по ID
      tags:
        - movies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Фильм успешно удалён
        '404':
          description: Фильм не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/{id}/actors:
    get:
      summary: Get actors of a movie
      description: Возвращает список актёров фильма
      tags:
        - actors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список актёров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Actor'
        '404':
          description: Фильм не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Movie:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
    NewMovie:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float
      required:
        - name
        - price
    Actor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string